remove unnecessary abstractions to simplify debug server code
