--- VSRAD.DebugServer/ConnectionFailedException.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace VSRAD.DebugServer
-{
-    internal sealed class ConnectionFailedException : Exception
-    {
-        public ConnectionFailedException()
-            : base("Client has been disconnected") { }
-
-        public ConnectionFailedException(string message)
-            : base(message) { }
-    }
-}
